satt sidevisning i useeffect istedet for onpressfunksjonen
