+++ src/App/InformasjonOmBedriftOgAnsatte/MineAnsatte/MineAnsatte.tsx
@@ -3,7 +3,6 @@ import './MineAnsatte.less';
 import { Undertittel } from 'nav-frontend-typografi';
 import { OrganisasjonsDetaljerContext } from '../../../OrganisasjonDetaljerProvider';
 import SideBytter from './BlaIAnsatte/BlaIAnsatte';
-import VisningAvSideBytter from './BlaIAnsatte/VisningAvPageTurner';
 
 interface Props {
     className?: string;
@@ -16,7 +15,8 @@ const MineAnsatte: FunctionComponent<Props> = props => {
             <Undertittel className={'mine-ansatte__systemtittel'} tabIndex={0}>
                 Opplysninger fra Aa-registeret
             </Undertittel>
-            <VisningAvSideBytter />
+
+            <SideBytter className={'sidebytter'} />
             <div tabIndex={0} className={'mine-ansatte__antall'}>
                 {mineAnsatte.arbeidsforholdoversikter.length} arbeidsforhold
             </div>
